website: support navigations
This CL includes prevSec/prevSecTitle/nextSec/nextSecTitle to specify the navigation in a section.

The feature is added but not activated everywhere due to a large mount of work need to be done.

Maybe documented and added to the .md file in the future. Check part2runtime/ch06sched/basic.md for example.

Updates #32